+++ packages/vendure-plugin-metrics/src/metrics.plugin.ts
@@ -8,7 +8,7 @@ import { MetricStrategy } from './api/metric-strategy';
 import { PLUGIN_INIT_OPTIONS } from './constants';
 import { RevenuePerProduct } from './api/metrics/revenue-per-product';
 import { AverageOrderValueMetric } from './api/metrics/average-order-value';
-import { SalesPerProductMetric } from './api/metrics/sales-per-product';
+import { UnitsSoldMetric } from './api/metrics/units-sold-metric';
 
 export interface MetricsPluginOptions {
   metrics: MetricStrategy<any>[];
@@ -31,7 +31,7 @@ export class MetricsPlugin {
     metrics: [
       new RevenuePerProduct(),
       new AverageOrderValueMetric(),
-      new SalesPerProductMetric(),
+      new UnitsSoldMetric(),
     ],
   };