+++ packages/vendure-plugin-metrics/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 - Added revenue (per variant) metric
 - Calculate with or without tax based on channel settings
-- Use number instead of currency formatting for Sales-per-product metric
+- Use number instead of currency formatting for Units Sold metric
 - Added a max cache age of 12 hours for metrics
 
 # 1.3.2 (2024-10-29)